Refactor game menu logic to Redux toolkit
