+++ src/components/game/game-info-block/prisoner/index.tsx
@@ -1,23 +1,17 @@
 import React from 'react'
-import { useSelector } from 'react-redux'
-
-import { InitialStateInterface } from '../../../../store/reducer'
+import { useAppSelector } from '../../../../store/hooks'
 
 const Prisoner: React.FC = function () {
-  const PrisonersArray = useSelector(
-    (state: InitialStateInterface) => state.prisoners
-  )
-  const prisonerIndex = useSelector(
-    (state: InitialStateInterface) => state.currentPrisonerId
-  )
+  const PrisonersArray = useAppSelector((state) => state.gameLogic.prisoners)
+  const prisonerIndex = useAppSelector((state) => state.gameLogic.currentPrisonerId)
   return (
     <div className="prisoner">
       <img
         className="prisoner__img"
-        src={PrisonersArray[prisonerIndex].prisonerImg}
-        alt={PrisonersArray[prisonerIndex].prisonerName}
+        src={PrisonersArray[prisonerIndex]?.prisonerImg}
+        alt={PrisonersArray[prisonerIndex]?.prisonerName}
       />
-      <p className="prisoner__name">{PrisonersArray[prisonerIndex].prisonerName}</p>
+      <p className="prisoner__name">{PrisonersArray[prisonerIndex]?.prisonerName}</p>
       {/* <p className="prisoner__index">{PrisonersArray[prisonerIndex].prisonerNumber}</p> */}
       {/* <p>Attempts {PrisonersArray[prisonerIndex].attempts}</p> */}
     </div>
